feat(gpu): signed and unsigned scalar mul
+ remove small scalar mul
+ move around signed tests_cases
