+++ backends/tfhe-cuda-backend/cuda/src/integer/shift_and_rotate.cuh
@@ -5,9 +5,9 @@
 #include "device.h"
 #include "integer.cuh"
 #include "integer.h"
-#include "multiplication.cuh"
 #include "pbs/programmable_bootstrap_classic.cuh"
 #include "pbs/programmable_bootstrap_multibit.cuh"
+#include "scalar_mul.cuh"
 #include "types/complex/operations.cuh"
 #include "utils/helper.cuh"
 #include "utils/kernel_dimensions.cuh"
@@ -157,7 +157,7 @@ __host__ void host_integer_radix_shift_and_rotate_kb_inplace(
   lwe_last_out = lwe_array;
   for (int i = bits_per_block - 2; i >= 0; i--) {
 
-    host_integer_small_scalar_mult_radix<Torus>(
+    host_integer_small_scalar_mul_radix<Torus>(
         stream, lwe_last_out, lwe_last_out, 2, big_lwe_dimension,
         num_radix_blocks);
