+++ tfhe/src/integer/server_key/radix_parallel/scalar_mul.rs
@@ -1,322 +1,10 @@
 use crate::integer::block_decomposition::{BlockDecomposer, DecomposableInto};
-use crate::integer::ciphertext::{IntegerRadixCiphertext, RadixCiphertext};
+use crate::integer::ciphertext::IntegerRadixCiphertext;
 use crate::integer::server_key::radix::scalar_mul::ScalarMultiplier;
-use crate::integer::server_key::CheckError;
 use crate::integer::ServerKey;
 use rayon::prelude::*;
 
 impl ServerKey {
-    /// Computes homomorphically a multiplication between a scalar and a ciphertext.
-    ///
-    /// This function computes the operation without checking if it exceeds the capacity of the
-    /// ciphertext.
-    ///
-    /// The result is returned as a new ciphertext.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use tfhe::integer::gen_keys_radix;
-    /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
-    ///
-    /// // We have 4 * 2 = 8 bits of message
-    /// let size = 4;
-    /// let (cks, sks) = gen_keys_radix(PARAM_MESSAGE_2_CARRY_2_KS_PBS, size);
-    ///
-    /// let msg = 30;
-    /// let scalar = 3;
-    ///
-    /// let ct = cks.encrypt(msg);
-    ///
-    /// // Compute homomorphically a scalar multiplication:
-    /// let ct_res = sks.unchecked_small_scalar_mul_parallelized(&ct, scalar);
-    ///
-    /// let clear: u64 = cks.decrypt(&ct_res);
-    /// assert_eq!(scalar * msg, clear);
-    /// ```
-    pub fn unchecked_small_scalar_mul_parallelized(
-        &self,
-        ctxt: &RadixCiphertext,
-        scalar: u64,
-    ) -> RadixCiphertext {
-        let mut ct_result = ctxt.clone();
-        self.unchecked_small_scalar_mul_assign_parallelized(&mut ct_result, scalar);
-        ct_result
-    }
-
-    pub fn unchecked_small_scalar_mul_assign_parallelized(
-        &self,
-        ctxt: &mut RadixCiphertext,
-        scalar: u64,
-    ) {
-        ctxt.blocks.par_iter_mut().for_each(|ct_i| {
-            self.key.unchecked_scalar_mul_assign(ct_i, scalar as u8);
-        });
-    }
-
-    /// Computes homomorphically a multiplication between a scalar and a ciphertext.
-    ///
-    /// If the operation can be performed, the result is returned in a new ciphertext.
-    /// Otherwise a [CheckError] is returned.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use tfhe::integer::gen_keys_radix;
-    /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
-    ///
-    /// // We have 4 * 2 = 8 bits of message
-    /// let size = 4;
-    /// let (cks, sks) = gen_keys_radix(PARAM_MESSAGE_2_CARRY_2_KS_PBS, size);
-    ///
-    /// let msg = 33;
-    /// let scalar = 3;
-    ///
-    /// let ct = cks.encrypt(msg);
-    ///
-    /// // Compute homomorphically a scalar multiplication:
-    /// let ct_res = sks.checked_small_scalar_mul_parallelized(&ct, scalar);
-    ///
-    /// match ct_res {
-    ///     Err(x) => panic!("{:?}", x),
-    ///     Ok(y) => {
-    ///         let clear: u64 = cks.decrypt(&y);
-    ///         assert_eq!(msg * scalar, clear);
-    ///     }
-    /// }
-    /// ```
-    pub fn checked_small_scalar_mul_parallelized(
-        &self,
-        ct: &RadixCiphertext,
-        scalar: u64,
-    ) -> Result<RadixCiphertext, CheckError> {
-        // If the ciphertext cannot be multiplied without exceeding the capacity of a ciphertext
-        self.is_small_scalar_mul_possible(ct, scalar)?;
-        Ok(self.unchecked_small_scalar_mul_parallelized(ct, scalar))
-    }
-
-    /// Computes homomorphically a multiplication between a scalar and a ciphertext.
-    ///
-    /// If the operation can be performed, the result is assigned to the ciphertext given
-    /// as parameter.
-    /// Otherwise a [CheckError] is returned.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use tfhe::integer::gen_keys_radix;
-    /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
-    ///
-    /// // We have 4 * 2 = 8 bits of message
-    /// let size = 4;
-    /// let (cks, sks) = gen_keys_radix(PARAM_MESSAGE_2_CARRY_2_KS_PBS, size);
-    ///
-    /// let msg = 33;
-    /// let scalar = 3;
-    ///
-    /// let mut ct = cks.encrypt(msg);
-    ///
-    /// // Compute homomorphically a scalar multiplication:
-    /// sks.checked_small_scalar_mul_assign_parallelized(&mut ct, scalar);
-    ///
-    /// let clear_res: u64 = cks.decrypt(&ct);
-    /// assert_eq!(clear_res, msg * scalar);
-    /// ```
-    pub fn checked_small_scalar_mul_assign_parallelized(
-        &self,
-        ct: &mut RadixCiphertext,
-        scalar: u64,
-    ) -> Result<(), CheckError> {
-        // If the ciphertext cannot be multiplied without exceeding the capacity of a ciphertext
-        self.is_small_scalar_mul_possible(ct, scalar)?;
-        self.unchecked_small_scalar_mul_assign_parallelized(ct, scalar);
-        Ok(())
-    }
-
-    /// Computes homomorphically a multiplication between a scalar and a ciphertext.
-    ///
-    /// `small` means the scalar value shall fit in a __shortint block__.
-    /// For example, if the parameters are PARAM_MESSAGE_2_CARRY_2_KS_PBS,
-    /// the scalar should fit in 2 bits.
-    ///
-    /// The result is returned as a new ciphertext.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use tfhe::integer::gen_keys_radix;
-    /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
-    ///
-    /// // We have 4 * 2 = 8 bits of message
-    /// let modulus = 1 << 8;
-    /// let size = 4;
-    /// let (cks, sks) = gen_keys_radix(PARAM_MESSAGE_2_CARRY_2_KS_PBS, size);
-    ///
-    /// let msg = 13;
-    /// let scalar = 3;
-    ///
-    /// let mut ct = cks.encrypt(msg);
-    ///
-    /// // Compute homomorphically a scalar multiplication:
-    /// let ct_res = sks.smart_small_scalar_mul_parallelized(&mut ct, scalar);
-    ///
-    /// // Decrypt:
-    /// let clear: u64 = cks.decrypt(&ct_res);
-    /// assert_eq!(msg * scalar % modulus, clear);
-    /// ```
-    pub fn smart_small_scalar_mul_parallelized(
-        &self,
-        ctxt: &mut RadixCiphertext,
-        scalar: u64,
-    ) -> RadixCiphertext {
-        if self.is_small_scalar_mul_possible(ctxt, scalar).is_err() {
-            self.full_propagate_parallelized(ctxt);
-        }
-        self.is_small_scalar_mul_possible(ctxt, scalar).unwrap();
-        self.unchecked_small_scalar_mul_parallelized(ctxt, scalar)
-    }
-
-    /// Computes homomorphically a multiplication between a scalar and a ciphertext.
-    ///
-    /// `small` means the scalar shall value fit in a __shortint block__.
-    /// For example, if the parameters are PARAM_MESSAGE_2_CARRY_2_KS_PBS,
-    /// the scalar should fit in 2 bits.
-    ///
-    /// The result is assigned to the input ciphertext
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use tfhe::integer::gen_keys_radix;
-    /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
-    ///
-    /// // We have 4 * 2 = 8 bits of message
-    /// let modulus = 1 << 8;
-    /// let size = 4;
-    /// let (cks, sks) = gen_keys_radix(PARAM_MESSAGE_2_CARRY_2_KS_PBS, size);
-    ///
-    /// let msg = 9;
-    /// let scalar = 3;
-    ///
-    /// let mut ct = cks.encrypt(msg);
-    ///
-    /// // Compute homomorphically a scalar multiplication:
-    /// sks.smart_small_scalar_mul_assign_parallelized(&mut ct, scalar);
-    ///
-    /// // Decrypt:
-    /// let clear: u64 = cks.decrypt(&ct);
-    /// assert_eq!(msg * scalar % modulus, clear);
-    /// ```
-    pub fn smart_small_scalar_mul_assign_parallelized(
-        &self,
-        ctxt: &mut RadixCiphertext,
-        scalar: u64,
-    ) {
-        if self.is_small_scalar_mul_possible(ctxt, scalar).is_err() {
-            self.full_propagate_parallelized(ctxt);
-        }
-        self.is_small_scalar_mul_possible(ctxt, scalar).unwrap();
-        self.unchecked_small_scalar_mul_assign_parallelized(ctxt, scalar);
-    }
-
-    /// Computes homomorphically a multiplication between a scalar and a ciphertext.
-    ///
-    /// `small` means the scalar value shall fit in a __shortint block__.
-    /// For example, if the parameters are PARAM_MESSAGE_2_CARRY_2_KS_PBS,
-    /// the scalar should fit in 2 bits.
-    ///
-    /// The result is returned as a new ciphertext.
-    ///
-    /// This function, like all "default" operations (i.e. not smart, checked or unchecked), will
-    /// check that the input ciphertexts block carries are empty and clears them if it's not the
-    /// case and the operation requires it. It outputs a ciphertext whose block carries are always
-    /// empty.
-    ///
-    /// This means that when using only "default" operations, a given operation (like add for
-    /// example) has always the same performance characteristics from one call to another and
-    /// guarantees correctness by pre-emptively clearing carries of output ciphertexts.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use tfhe::integer::gen_keys_radix;
-    /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
-    ///
-    /// // We have 4 * 2 = 8 bits of message
-    /// let modulus = 1 << 8;
-    /// let size = 4;
-    /// let (cks, sks) = gen_keys_radix(PARAM_MESSAGE_2_CARRY_2_KS_PBS, size);
-    ///
-    /// let msg = 13;
-    /// let scalar = 3;
-    ///
-    /// let mut ct = cks.encrypt(msg);
-    ///
-    /// // Compute homomorphically a scalar multiplication:
-    /// let ct_res = sks.small_scalar_mul_parallelized(&mut ct, scalar);
-    ///
-    /// // Decrypt:
-    /// let clear: u64 = cks.decrypt(&ct_res);
-    /// assert_eq!(msg * scalar % modulus, clear);
-    /// ```
-    pub fn small_scalar_mul_parallelized(
-        &self,
-        ctxt: &RadixCiphertext,
-        scalar: u64,
-    ) -> RadixCiphertext {
-        let mut ct_res = ctxt.clone();
-        self.small_scalar_mul_assign_parallelized(&mut ct_res, scalar);
-        ct_res
-    }
-
-    /// Computes homomorphically a multiplication between a scalar and a ciphertext.
-    ///
-    /// `small` means the scalar shall value fit in a __shortint block__.
-    /// For example, if the parameters are PARAM_MESSAGE_2_CARRY_2_KS_PBS,
-    /// the scalar should fit in 2 bits.
-    ///
-    /// The result is assigned to the input ciphertext
-    ///
-    /// This function, like all "default" operations (i.e. not smart, checked or unchecked), will
-    /// check that the input ciphertexts block carries are empty and clears them if it's not the
-    /// case and the operation requires it. It outputs a ciphertext whose block carries are always
-    /// empty.
-    ///
-    /// This means that when using only "default" operations, a given operation (like add for
-    /// example) has always the same performance characteristics from one call to another and
-    /// guarantees correctness by pre-emptively clearing carries of output ciphertexts.
-    ///
-    /// # Example
-    ///
-    /// ```rust
-    /// use tfhe::integer::gen_keys_radix;
-    /// use tfhe::shortint::parameters::PARAM_MESSAGE_2_CARRY_2_KS_PBS;
-    ///
-    /// // We have 4 * 2 = 8 bits of message
-    /// let modulus = 1 << 8;
-    /// let size = 4;
-    /// let (cks, sks) = gen_keys_radix(PARAM_MESSAGE_2_CARRY_2_KS_PBS, size);
-    ///
-    /// let msg = 9;
-    /// let scalar = 3;
-    ///
-    /// let mut ct = cks.encrypt(msg);
-    ///
-    /// // Compute homomorphically a scalar multiplication:
-    /// sks.small_scalar_mul_assign_parallelized(&mut ct, scalar);
-    ///
-    /// // Decrypt:
-    /// let clear: u64 = cks.decrypt(&ct);
-    /// assert_eq!(msg * scalar % modulus, clear);
-    /// ```
-    pub fn small_scalar_mul_assign_parallelized(&self, ctxt: &mut RadixCiphertext, scalar: u64) {
-        if !ctxt.block_carries_are_empty() {
-            self.full_propagate_parallelized(ctxt);
-        }
-        self.unchecked_small_scalar_mul_assign_parallelized(ctxt, scalar);
-        self.full_propagate_parallelized(ctxt);
-    }
-
     pub fn unchecked_scalar_mul_parallelized<T, Scalar>(&self, ct: &T, scalar: Scalar) -> T
     where
         T: IntegerRadixCiphertext,
