--- tfhe/src/integer/server_key/radix_parallel/tests_cases_signed.rs
@@ -1,1933 +1,0 @@
-use crate::integer::keycache::KEY_CACHE;
-use crate::integer::server_key::radix_parallel::tests_cases_unsigned::FunctionExecutor;
-use crate::integer::{BooleanBlock, IntegerKeyKind, RadixClientKey, SignedRadixCiphertext};
-use crate::shortint::ciphertext::NoiseLevel;
-use crate::shortint::PBSParameters;
-use itertools::izip;
-use rand::prelude::ThreadRng;
-use rand::Rng;
-use std::sync::Arc;
-
-/// Number of loop iteration within randomized tests
-#[cfg(not(tarpaulin))]
-pub(crate) const NB_TESTS: usize = 30;
-/// Smaller number of loop iteration within randomized test,
-/// meant for test where the function tested is more expensive
-#[cfg(not(tarpaulin))]
-pub(crate) const NB_TESTS_SMALLER: usize = 10;
-#[cfg(not(tarpaulin))]
-pub(crate) const NB_TESTS_UNCHECKED: usize = NB_TESTS;
-
-// Use lower numbers for coverage to ensure fast tests to counter balance slowdown due to code
-// instrumentation
-#[cfg(tarpaulin)]
-pub(crate) const NB_TESTS: usize = 1;
-#[cfg(tarpaulin)]
-pub(crate) const NB_TESTS_SMALLER: usize = 1;
-/// Unchecked test cases needs a minimum number of tests of 4 in order to provide guarantees.
-#[cfg(tarpaulin)]
-pub(crate) const NB_TESTS_UNCHECKED: usize = 4;
-
-#[cfg(not(tarpaulin))]
-pub(crate) const NB_CTXT: usize = 4;
-#[cfg(tarpaulin)]
-pub(crate) const NB_CTXT: usize = 2;
-
-//================================================================================
-//     Helper functions
-//================================================================================
-
-pub(crate) fn signed_add_under_modulus(lhs: i64, rhs: i64, modulus: i64) -> i64 {
-    signed_overflowing_add_under_modulus(lhs, rhs, modulus).0
-}
-
-// Adds two signed number modulo the given modulus
-//
-// This is to 'simulate' i8, i16, ixy using i64 integers
-//
-// lhs and rhs must be in [-modulus..modulus[
-pub(crate) fn signed_overflowing_add_under_modulus(
-    lhs: i64,
-    rhs: i64,
-    modulus: i64,
-) -> (i64, bool) {
-    assert!(modulus > 0);
-    assert!((-modulus..modulus).contains(&lhs));
-
-    // The code below requires rhs and lhs to be in range -modulus..modulus
-    // in scalar tests, rhs may exceed modulus
-    // so we truncate it (is the fhe ops does)
-    let (mut res, mut overflowed) = if (-modulus..modulus).contains(&rhs) {
-        (lhs + rhs, false)
-    } else {
-        // 2*modulus to get all the bits
-        (lhs + (rhs % (2 * modulus)), true)
-    };
-
-    if res < -modulus {
-        // rem_euclid(modulus) would also work
-        res = modulus + (res - -modulus);
-        overflowed = true;
-    } else if res > modulus - 1 {
-        res = -modulus + (res - modulus);
-        overflowed = true;
-    }
-    (res, overflowed)
-}
-
-pub(crate) fn signed_neg_under_modulus(lhs: i64, modulus: i64) -> i64 {
-    assert!(modulus > 0);
-    let mut res = -lhs;
-    if res < -modulus {
-        // rem_euclid(modulus) would also work
-        res = modulus + (res - -modulus);
-    } else if res > modulus - 1 {
-        res = -modulus + (res - modulus);
-    }
-    res
-}
-
-// Subs two signed number modulo the given modulus
-//
-// This is to 'simulate' i8, i16, ixy using i64 integers
-//
-// lhs and rhs must be in [-modulus..modulus[
-pub(crate) fn signed_sub_under_modulus(lhs: i64, rhs: i64, modulus: i64) -> i64 {
-    signed_overflowing_sub_under_modulus(lhs, rhs, modulus).0
-}
-
-pub(crate) fn signed_overflowing_sub_under_modulus(
-    lhs: i64,
-    rhs: i64,
-    modulus: i64,
-) -> (i64, bool) {
-    // Technically we should be able to call overflowing_add_under_modulus(lhs, -rhs, ...)
-    // but due to -rhs being a 'special case' when rhs == -modulus, we have to
-    // so the impl here
-    assert!(modulus > 0);
-    assert!((-modulus..modulus).contains(&lhs));
-
-    // The code below requires rhs and lhs to be in range -modulus..modulus
-    // in scalar tests, rhs may exceed modulus
-    // so we truncate it (is the fhe ops does)
-    let (mut res, mut overflowed) = if (-modulus..modulus).contains(&rhs) {
-        (lhs - rhs, false)
-    } else {
-        // 2*modulus to get all the bits
-        (lhs - (rhs % (2 * modulus)), true)
-    };
-
-    if res < -modulus {
-        // rem_euclid(modulus) would also work
-        res = modulus + (res - -modulus);
-        overflowed = true;
-    } else if res > modulus - 1 {
-        res = -modulus + (res - modulus);
-        overflowed = true;
-    }
-    (res, overflowed)
-}
-
-pub(crate) fn signed_mul_under_modulus(lhs: i64, rhs: i64, modulus: i64) -> i64 {
-    assert!(modulus > 0);
-    overflowing_mul_under_modulus(lhs, rhs, modulus).0
-}
-
-pub(crate) fn overflowing_mul_under_modulus(lhs: i64, rhs: i64, modulus: i64) -> (i64, bool) {
-    let (mut res, mut overflowed) = lhs.overflowing_mul(rhs);
-    overflowed |= res < -modulus || res >= modulus;
-    res %= modulus * 2;
-    if res < -modulus {
-        // rem_euclid(modulus) would also work
-        res = modulus + (res - -modulus);
-    } else if res > modulus - 1 {
-        res = -modulus + (res - modulus);
-    }
-
-    (res, overflowed)
-}
-
-pub(crate) fn absolute_value_under_modulus(lhs: i64, modulus: i64) -> i64 {
-    if lhs < 0 {
-        signed_neg_under_modulus(lhs, modulus)
-    } else {
-        lhs
-    }
-}
-
-pub(crate) fn signed_left_shift_under_modulus(lhs: i64, rhs: u32, modulus: i64) -> i64 {
-    signed_mul_under_modulus(lhs, 1 << rhs, modulus)
-}
-
-pub(crate) fn signed_right_shift_under_modulus(lhs: i64, rhs: u32, _modulus: i64) -> i64 {
-    lhs >> rhs
-}
-
-pub(crate) fn signed_div_under_modulus(lhs: i64, rhs: i64, modulus: i64) -> i64 {
-    // in signed integers, -modulus can be represented, but +modulus cannot
-    // thus, when dividing: -128 / -1 = 128 the results overflows to -128
-    assert!(modulus > 0);
-    let mut res = lhs / rhs;
-    if res < -modulus {
-        // rem_euclid(modulus) would also work
-        res = modulus + (res - -modulus);
-    } else if res > modulus - 1 {
-        res = -modulus + (res - modulus);
-    }
-    res
-}
-
-pub(crate) fn signed_rem_under_modulus(lhs: i64, rhs: i64, modulus: i64) -> i64 {
-    assert!(modulus > 0);
-    let q = signed_div_under_modulus(lhs, rhs, modulus);
-    let q_times_rhs = signed_mul_under_modulus(q, rhs, modulus);
-    signed_sub_under_modulus(lhs, q_times_rhs, modulus)
-}
-
-pub(crate) fn signed_div_rem_floor_under_modulus(lhs: i64, rhs: i64, modulus: i64) -> (i64, i64) {
-    let mut q = signed_div_under_modulus(lhs, rhs, modulus);
-    let mut r = signed_rem_under_modulus(lhs, rhs, modulus);
-
-    if (r != 0) && ((r < 0) != (rhs < 0)) {
-        q = signed_sub_under_modulus(q, 1, modulus);
-        r = signed_add_under_modulus(r, rhs, modulus);
-    }
-
-    (q, r)
-}
-
-/// helper function to do a rotate left when the type used to store
-/// the value is bigger than the actual intended bit size
-pub(crate) fn rotate_left_helper(value: i64, n: u32, actual_bit_size: u32) -> i64 {
-    // We start with:
-    // [0000000000000|xxxx]
-    // 64           b    0
-    //
-    // rotated will be
-    // [0000000000xx|xx00]
-    // 64           b    0
-    let n = n % actual_bit_size;
-    let mask = 1i64.wrapping_shl(actual_bit_size) - 1;
-    let shifted_mask = mask.wrapping_shl(n) & !mask;
-
-    // Value maybe be negative and so, have its msb
-    // set to one, so use mask to only keep the part that interest
-    // us
-    let rotated = (value & mask).rotate_left(n);
-
-    let tmp = (rotated & mask) | ((rotated & shifted_mask) >> actual_bit_size);
-    // If the sign bit after rotation is one,
-    // then all bits above it needs to be one
-    let new_sign_bit = (tmp >> (actual_bit_size - 1)) & 1;
-    let mut pad = -new_sign_bit;
-    pad <<= actual_bit_size; // only bits above actual_bit_size should be set
-
-    pad | tmp
-}
-
-/// helper function to do a rotate right when the type used to store
-/// the value is bigger than the actual intended bit size
-pub(crate) fn rotate_right_helper(value: i64, n: u32, actual_bit_size: u32) -> i64 {
-    // We start with:
-    // [yyyyyyyyyyyy|xxxx]
-    // 64           b    0
-    // where xs are bits that we are interested in
-    // and ys are either 0 or 1 depending on if value is positive
-    //
-    // mask: [yyyyyyyyyyyy|mmmm]
-    // shifted_ mask: [mmyyyyyyyyyy|0000]
-    //
-    // rotated will be
-    // [xxyyyyyyyyyy|00xx]
-    // 64           b    0
-    //
-    // To get the 'cycled' bits where they should be,
-    // we get them using a mask then shift
-    let n = n % actual_bit_size;
-    let mask = 1i64.wrapping_shl(actual_bit_size) - 1;
-    // shifted mask only needs the bits that cycled
-    let shifted_mask = mask.rotate_right(n) & !mask;
-
-    // Value maybe be negative and so, have its msb
-    // set to one, so use mask to only keep the part that interest
-    // us
-    let rotated = (value & mask).rotate_right(n);
-
-    let tmp = (rotated & mask) | ((rotated & shifted_mask) >> (u64::BITS - actual_bit_size));
-    // If the sign bit after rotation is one,
-    // then all bits above it needs to be one
-    let new_sign_bit = (tmp >> (actual_bit_size - 1)) & 1;
-    let mut pad = -new_sign_bit;
-    pad <<= actual_bit_size; // only bits above actual_bit_size should be set
-
-    pad | tmp
-}
-
-/// Returns an array filled with random values such that:
-/// - the first half contains values in [0..modulus[
-/// - the second half contains values in [-modulus..0]
-pub(crate) fn random_signed_value_under_modulus<const N: usize>(
-    rng: &mut rand::prelude::ThreadRng,
-    modulus: i64,
-) -> [i64; N] {
-    assert!(modulus > 0);
-
-    let mut values = [0i64; N];
-
-    for value in &mut values[..N / 2] {
-        *value = rng.gen_range(0..modulus);
-    }
-
-    for value in &mut values[N / 2..] {
-        *value = rng.gen_range(-modulus..=0);
-    }
-
-    values
-}
-
-/// Returns an array filled with random values such that:
-/// - the first half contains values in ]0..modulus[
-/// - the second half contains values in [-modulus..0[
-pub(crate) fn random_non_zero_signed_value_under_modulus<const N: usize>(
-    rng: &mut rand::prelude::ThreadRng,
-    modulus: i64,
-) -> [i64; N] {
-    assert!(modulus > 0);
-
-    let mut values = [0i64; N];
-
-    for value in &mut values[..N / 2] {
-        *value = rng.gen_range(1..modulus);
-    }
-
-    for value in &mut values[N / 2..] {
-        *value = rng.gen_range(-modulus..0);
-    }
-
-    values
-}
-
-/// Returns an iterator that yields pairs of i64 values in range `-modulus..modulus`
-/// such that there is at least one pair of (P, P), (P, N), (N, N) (N, P)
-/// where P means value >=0 and N means <= 0
-pub(crate) fn create_iterator_of_signed_random_pairs<const N: usize>(
-    rng: &mut rand::prelude::ThreadRng,
-    modulus: i64,
-) -> impl Iterator<Item = (i64, i64)> {
-    assert!(N >= 4, "N must be at least 4 to uphold the guarantee");
-    let mut lhs_values = [0i64; N];
-    let mut rhs_values = [0i64; N];
-
-    lhs_values[0] = rng.gen_range(0..modulus);
-    rhs_values[0] = rng.gen_range(0..modulus);
-
-    lhs_values[1] = rng.gen_range(0..modulus);
-    rhs_values[1] = rng.gen_range(-modulus..=0);
-
-    lhs_values[2] = rng.gen_range(-modulus..=0);
-    rhs_values[2] = rng.gen_range(-modulus..=0);
-
-    lhs_values[3] = rng.gen_range(-modulus..=0);
-    rhs_values[3] = rng.gen_range(0..modulus);
-
-    for i in 4..N {
-        lhs_values[i] = rng.gen_range(-modulus..modulus);
-        rhs_values[i] = rng.gen_range(-modulus..modulus);
-    }
-
-    izip!(lhs_values, rhs_values)
-}
-
-pub(crate) fn random_non_zero_value(rng: &mut ThreadRng, modulus: i64) -> i64 {
-    loop {
-        let value = rng.gen::<i64>() % modulus;
-        if value != 0 {
-            break value;
-        }
-    }
-}
-
-// Signed tests
-
-pub(crate) fn signed_unchecked_add_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let sks = Arc::new(sks);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    // check some overflow behaviour
-    let overflowing_values = [
-        (-modulus, -1, modulus - 1),
-        (modulus - 1, 1, -modulus),
-        (-modulus, -2, modulus - 2),
-        (modulus - 2, 2, -modulus),
-    ];
-    for (clear_0, clear_1, expected_clear) in overflowing_values {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = signed_add_under_modulus(clear_0, clear_1, modulus);
-        assert_eq!(clear_res, dec_res);
-        assert_eq!(clear_res, expected_clear);
-    }
-
-    for (clear_0, clear_1) in
-        create_iterator_of_signed_random_pairs::<NB_TESTS_UNCHECKED>(&mut rng, modulus)
-    {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = signed_add_under_modulus(clear_0, clear_1, modulus);
-        assert_eq!(clear_res, dec_res);
-    }
-}
-
-pub(crate) fn signed_default_add_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    sks.set_deterministic_pbs_execution(true);
-    let sks = Arc::new(sks);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    let mut clear;
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let mut ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let tmp_ct = executor.execute((&ctxt_0, &ctxt_1));
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp_ct);
-
-        clear = signed_add_under_modulus(clear_0, clear_1, modulus);
-
-        // println!("clear_0 = {}, clear_1 = {}", clear_0, clear_1);
-        // add multiple times to raise the degree
-        for _ in 0..NB_TESTS_SMALLER {
-            ct_res = executor.execute((&ct_res, &ctxt_0));
-            assert!(ct_res.block_carries_are_empty());
-            clear = signed_add_under_modulus(clear, clear_0, modulus);
-
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-
-            // println!("clear = {}, dec_res = {}", clear, dec_res);
-            assert_eq!(clear, dec_res);
-        }
-    }
-}
-
-pub(crate) fn signed_smart_add_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a mut SignedRadixCiphertext, &'a mut SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let sks = Arc::new(sks);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    let mut clear;
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear_0 = rng.gen_range(-modulus..modulus);
-        let clear_1 = rng.gen_range(-modulus..modulus);
-
-        let mut ctxt_0 = cks.encrypt_signed(clear_0);
-        let mut ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let mut ct_res = executor.execute((&mut ctxt_0, &mut ctxt_1));
-        clear = signed_add_under_modulus(clear_0, clear_1, modulus);
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        assert_eq!(clear, dec_res);
-
-        // add multiple times to raise the degree
-        for _ in 0..NB_TESTS_SMALLER {
-            ct_res = executor.execute((&mut ct_res, &mut ctxt_0));
-            clear = signed_add_under_modulus(clear, clear_0, modulus);
-
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-
-            // println!("clear = {}, dec_res = {}", clear, dec_res);
-            assert_eq!(clear, dec_res);
-        }
-    }
-}
-
-pub(crate) fn signed_unchecked_sub_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let sks = Arc::new(sks);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    // check some overflow behaviour
-    let overflowing_values = [
-        (-modulus, 1, modulus - 1),
-        (modulus - 1, -1, -modulus),
-        (-modulus, 2, modulus - 2),
-        (modulus - 2, -2, -modulus),
-    ];
-    for (clear_0, clear_1, expected_clear) in overflowing_values {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = signed_sub_under_modulus(clear_0, clear_1, modulus);
-        assert_eq!(clear_res, dec_res);
-        assert_eq!(clear_res, expected_clear);
-    }
-
-    for (clear_0, clear_1) in
-        create_iterator_of_signed_random_pairs::<NB_TESTS_UNCHECKED>(&mut rng, modulus)
-    {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = signed_sub_under_modulus(clear_0, clear_1, modulus);
-        assert_eq!(clear_res, dec_res);
-    }
-}
-
-pub(crate) fn signed_default_sub_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    sks.set_deterministic_pbs_execution(true);
-    let sks = Arc::new(sks);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    let mut rng = rand::thread_rng();
-
-    // message_modulus^vec_length
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    let mut clear;
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let mut ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let tmp_ct = executor.execute((&ctxt_0, &ctxt_1));
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp_ct);
-
-        clear = signed_sub_under_modulus(clear_0, clear_1, modulus);
-
-        // sub multiple times to raise the degree
-        for _ in 0..NB_TESTS_SMALLER {
-            ct_res = executor.execute((&ct_res, &ctxt_0));
-            assert!(ct_res.block_carries_are_empty());
-            clear = signed_sub_under_modulus(clear, clear_0, modulus);
-
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-
-            // println!("clear = {}, dec_res = {}", clear, dec_res);
-            assert_eq!(clear, dec_res);
-        }
-    }
-}
-
-pub(crate) fn signed_unchecked_neg_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<&'a SignedRadixCiphertext, SignedRadixCiphertext>,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let ctxt_zero = sks.create_trivial_radix(0i64, NB_CTXT);
-    let sks = Arc::new(sks);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    // -modulus is a special case, its negation cannot be
-    // represented. rust by default returns -modulus
-    // (which is what two complement result in)
-    {
-        let clear = -modulus;
-        let ctxt = cks.encrypt_signed(clear);
-
-        let ct_res = executor.execute(&ctxt);
-
-        let dec: i64 = cks.decrypt_signed(&ct_res);
-        let clear_result = signed_neg_under_modulus(clear, modulus);
-
-        assert_eq!(clear_result, dec);
-        assert_eq!(clear_result, -modulus);
-    }
-
-    for (clear_0, _) in
-        create_iterator_of_signed_random_pairs::<NB_TESTS_UNCHECKED>(&mut rng, modulus)
-    {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let ct_res = executor.execute(&ctxt_0);
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = signed_neg_under_modulus(clear_0, modulus);
-        assert_eq!(clear_res, dec_res);
-    }
-
-    // negation of trivial 0
-    {
-        let ct_res = executor.execute(&ctxt_zero);
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        assert_eq!(0, dec_res);
-    }
-}
-
-pub(crate) fn signed_smart_neg_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<&'a mut SignedRadixCiphertext, SignedRadixCiphertext>,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear = rng.gen::<i64>() % modulus;
-
-        let mut ctxt = cks.encrypt_signed(clear);
-
-        let mut ct_res = executor.execute(&mut ctxt);
-        let mut clear_res = signed_neg_under_modulus(clear, modulus);
-        let dec: i64 = cks.decrypt_signed(&ct_res);
-        assert_eq!(clear_res, dec);
-
-        for _ in 0..NB_TESTS_SMALLER {
-            ct_res = executor.execute(&mut ct_res);
-            clear_res = signed_neg_under_modulus(clear_res, modulus);
-
-            let dec: i64 = cks.decrypt_signed(&ct_res);
-            println!("clear_res: {clear_res}, dec : {dec}");
-            assert_eq!(clear_res, dec);
-        }
-    }
-}
-
-pub(crate) fn signed_default_neg_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<&'a SignedRadixCiphertext, SignedRadixCiphertext>,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    sks.set_deterministic_pbs_execution(true);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    // -modulus is a special case, its negation cannot be
-    // represented. rust by default returns -modulus
-    // (which is what two complement result in)
-    {
-        let clear = -modulus;
-        let ctxt = cks.encrypt_signed(clear);
-
-        let ct_res = executor.execute(&ctxt);
-        let tmp = executor.execute(&ctxt);
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp);
-
-        let dec: i64 = cks.decrypt_signed(&ct_res);
-        let clear_result = signed_neg_under_modulus(clear, modulus);
-
-        assert_eq!(clear_result, dec);
-    }
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear = rng.gen::<i64>() % modulus;
-
-        let ctxt = cks.encrypt_signed(clear);
-
-        let ct_res = executor.execute(&ctxt);
-        let tmp = executor.execute(&ctxt);
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp);
-
-        let dec: i64 = cks.decrypt_signed(&ct_res);
-        let clear_result = signed_neg_under_modulus(clear, modulus);
-
-        assert_eq!(clear_result, dec);
-    }
-}
-
-pub(crate) fn signed_unchecked_mul_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    for (clear_0, clear_1) in
-        create_iterator_of_signed_random_pairs::<NB_TESTS_UNCHECKED>(&mut rng, modulus)
-    {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = signed_mul_under_modulus(clear_0, clear_1, modulus);
-        assert_eq!(clear_res, dec_res);
-    }
-}
-
-pub(crate) fn signed_default_mul_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    sks.set_deterministic_pbs_execution(true);
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    // message_modulus^vec_length
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    let mut clear;
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let mut ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let tmp_ct = executor.execute((&ctxt_0, &ctxt_1));
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp_ct);
-
-        clear = signed_mul_under_modulus(clear_0, clear_1, modulus);
-
-        // mul multiple times to raise the degree
-        for _ in 0..NB_TESTS_SMALLER {
-            ct_res = executor.execute((&ct_res, &ctxt_0));
-            assert!(ct_res.block_carries_are_empty());
-            clear = signed_mul_under_modulus(clear, clear_0, modulus);
-
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-
-            // println!("clear = {}, dec_res = {}", clear, dec_res);
-            assert_eq!(clear, dec_res);
-        }
-    }
-}
-
-pub(crate) fn signed_unchecked_scalar_add_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<(&'a SignedRadixCiphertext, i64), SignedRadixCiphertext>,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let sks = Arc::new(sks);
-    let cks = RadixClientKey::from((
-        cks,
-        crate::integer::server_key::radix_parallel::tests_cases_unsigned::NB_CTXT,
-    ));
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    // check some overflow behaviour
-    let overflowing_values = [
-        (-modulus, -1, modulus - 1),
-        (modulus - 1, 1, -modulus),
-        (-modulus, -2, modulus - 2),
-        (modulus - 2, 2, -modulus),
-    ];
-    for (clear_0, clear_1, expected_clear) in overflowing_values {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ct_res = executor.execute((&ctxt_0, clear_1));
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = signed_add_under_modulus(clear_0, clear_1, modulus);
-        assert_eq!(clear_res, dec_res);
-        assert_eq!(clear_res, expected_clear);
-    }
-
-    for _ in 0..NB_TESTS {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let ct_res = executor.execute((&ctxt_0, clear_1));
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = signed_add_under_modulus(clear_0, clear_1, modulus);
-        assert_eq!(clear_res, dec_res);
-    }
-}
-
-pub(crate) fn signed_default_scalar_add_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<(&'a SignedRadixCiphertext, i64), SignedRadixCiphertext>,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    sks.set_deterministic_pbs_execution(true);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    // message_modulus^vec_length
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    let mut clear;
-
-    let mut rng = rand::thread_rng();
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let mut ct_res = executor.execute((&ctxt_0, clear_1));
-        assert!(ct_res.block_carries_are_empty());
-
-        clear = signed_add_under_modulus(clear_0, clear_1, modulus);
-
-        // add multiple times to raise the degree
-        for _ in 0..NB_TESTS_SMALLER {
-            let tmp = executor.execute((&ct_res, clear_1));
-            ct_res = executor.execute((&ct_res, clear_1));
-            assert!(ct_res.block_carries_are_empty());
-            assert_eq!(ct_res, tmp);
-            clear = signed_add_under_modulus(clear, clear_1, modulus);
-
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-            assert_eq!(clear, dec_res);
-        }
-    }
-}
-
-pub(crate) fn signed_default_overflowing_scalar_add_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, i64),
-        (SignedRadixCiphertext, BooleanBlock),
-    >,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    sks.set_deterministic_pbs_execution(true);
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    // message_modulus^vec_length
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks.clone());
-
-    let hardcoded_values = [
-        (-modulus, -1),
-        (modulus - 1, 1),
-        (-1, -modulus),
-        (1, modulus - 1),
-    ];
-    for (clear_0, clear_1) in hardcoded_values {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let (ct_res, result_overflowed) = executor.execute((&ctxt_0, clear_1));
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_add_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-        let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for add, for ({clear_0} + {clear_1}) % {modulus} \
-             expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_add for ({clear_0} + {clear_1}) % {modulus} \
-             expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert_eq!(result_overflowed.0.degree.get(), 1);
-        assert_eq!(result_overflowed.0.noise_level(), NoiseLevel::NOMINAL);
-    }
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let (ct_res, result_overflowed) = executor.execute((&ctxt_0, clear_1));
-        let (tmp_ct, tmp_o) = sks.signed_overflowing_scalar_add_parallelized(&ctxt_0, clear_1);
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp_ct, "Failed determinism check");
-        assert_eq!(tmp_o, result_overflowed, "Failed determinism check");
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_add_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-        let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for add, for ({clear_0} + {clear_1}) % {modulus} \
-             expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_add for ({clear_0} + {clear_1}) % {modulus} \
-             expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert_eq!(result_overflowed.0.degree.get(), 1);
-        assert_eq!(result_overflowed.0.noise_level(), NoiseLevel::NOMINAL);
-
-        for _ in 0..NB_TESTS_SMALLER {
-            // Add non zero scalar to have non clean ciphertexts
-            let clear_2 = random_non_zero_value(&mut rng, modulus);
-            let clear_rhs = random_non_zero_value(&mut rng, modulus);
-
-            let ctxt_0 = sks.unchecked_scalar_add(&ctxt_0, clear_2);
-            let (clear_lhs, _) = signed_overflowing_add_under_modulus(clear_0, clear_2, modulus);
-            let d0: i64 = cks.decrypt_signed(&ctxt_0);
-            assert_eq!(d0, clear_lhs, "Failed sanity decryption check");
-
-            let (ct_res, result_overflowed) = executor.execute((&ctxt_0, clear_rhs));
-            assert!(ct_res.block_carries_are_empty());
-            let (expected_result, expected_overflowed) =
-                signed_overflowing_add_under_modulus(clear_lhs, clear_rhs, modulus);
-
-            let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-            let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-            assert_eq!(
-                decrypted_result, expected_result,
-                "Invalid result for add, for ({clear_lhs} + {clear_rhs}) % {modulus} \
-                expected {expected_result}, got {decrypted_result}"
-            );
-            assert_eq!(
-                decrypted_overflowed,
-                expected_overflowed,
-                "Invalid overflow flag result for overflowing_add, for ({clear_lhs} + {clear_rhs}) % {modulus} \
-                expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-            );
-            assert_eq!(result_overflowed.0.degree.get(), 1);
-            assert_eq!(result_overflowed.0.noise_level(), NoiseLevel::NOMINAL);
-        }
-    }
-
-    // Test with trivial inputs
-    for _ in 0..4 {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let a: SignedRadixCiphertext = sks.create_trivial_radix(clear_0, NB_CTXT);
-
-        let (encrypted_result, encrypted_overflow) = executor.execute((&a, clear_1));
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_add_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&encrypted_result);
-        let decrypted_overflowed = cks.decrypt_bool(&encrypted_overflow);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for add, for ({clear_0} + {clear_1}) % {modulus} \
-                expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_add, for ({clear_0} + {clear_1}) % {modulus} \
-                expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert_eq!(encrypted_overflow.0.degree.get(), 1);
-        assert_eq!(encrypted_overflow.0.noise_level(), NoiseLevel::ZERO);
-    }
-
-    // Test with scalar that is bigger than ciphertext modulus
-    for _ in 0..2 {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen_range(modulus..=i64::MAX);
-
-        let a = cks.encrypt_signed(clear_0);
-
-        let (encrypted_result, encrypted_overflow) = executor.execute((&a, clear_1));
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_add_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&encrypted_result);
-        let decrypted_overflowed = cks.decrypt_bool(&encrypted_overflow);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for overflowing_add, for ({clear_0} + {clear_1}) % {modulus} \
-                expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_add, for ({clear_0} + {clear_1}) % {modulus} \
-                expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert!(decrypted_overflowed); // Actually we know its an overflow case
-        assert_eq!(encrypted_overflow.0.degree.get(), 1);
-        assert_eq!(encrypted_overflow.0.noise_level(), NoiseLevel::ZERO);
-    }
-}
-
-pub(crate) fn signed_unchecked_scalar_sub_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<(&'a SignedRadixCiphertext, i64), SignedRadixCiphertext>,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    // check some overflow behaviour
-    let overflowing_values = [
-        (-modulus, 1, modulus - 1),
-        (modulus - 1, -1, -modulus),
-        (-modulus, 2, modulus - 2),
-        (modulus - 2, -2, -modulus),
-    ];
-    for (clear_0, clear_1, expected_clear) in overflowing_values {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ct_res = executor.execute((&ctxt_0, clear_1));
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = signed_sub_under_modulus(clear_0, clear_1, modulus);
-        assert_eq!(clear_res, dec_res);
-        assert_eq!(clear_res, expected_clear);
-    }
-
-    for _ in 0..NB_TESTS {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let ct_res = executor.execute((&ctxt_0, clear_1));
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = signed_sub_under_modulus(clear_0, clear_1, modulus);
-        assert_eq!(clear_res, dec_res);
-    }
-}
-
-pub(crate) fn signed_default_overflowing_scalar_sub_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, i64),
-        (SignedRadixCiphertext, BooleanBlock),
-    >,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    sks.set_deterministic_pbs_execution(true);
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    // message_modulus^vec_length
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks.clone());
-
-    let hardcoded_values = [
-        (-modulus, 1),
-        (modulus - 1, -1),
-        (1, -modulus),
-        (-1, modulus - 1),
-    ];
-    for (clear_0, clear_1) in hardcoded_values {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let (ct_res, result_overflowed) = executor.execute((&ctxt_0, clear_1));
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_sub_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-        let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for overflowing_sub, for ({clear_0} - {clear_1}) % {modulus} \
-             expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_sub for ({clear_0} - {clear_1}) % {modulus} \
-             expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert_eq!(result_overflowed.0.degree.get(), 1);
-        assert_eq!(result_overflowed.0.noise_level(), NoiseLevel::NOMINAL);
-    }
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let (ct_res, result_overflowed) = executor.execute((&ctxt_0, clear_1));
-        let (tmp_ct, tmp_o) = executor.execute((&ctxt_0, clear_1));
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp_ct, "Failed determinism check");
-        assert_eq!(tmp_o, result_overflowed, "Failed determinism check");
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_sub_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-        let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for sub, for ({clear_0} - {clear_1}) % {modulus} \
-             expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_sub for ({clear_0} - {clear_1}) % {modulus} \
-             expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert_eq!(result_overflowed.0.degree.get(), 1);
-        assert_eq!(result_overflowed.0.noise_level(), NoiseLevel::NOMINAL);
-
-        for _ in 0..NB_TESTS_SMALLER {
-            // Add non zero scalar to have non clean ciphertexts
-            let clear_2 = random_non_zero_value(&mut rng, modulus);
-            let clear_rhs = random_non_zero_value(&mut rng, modulus);
-
-            let ctxt_0 = sks.unchecked_scalar_add(&ctxt_0, clear_2);
-            let (clear_lhs, _) = signed_overflowing_add_under_modulus(clear_0, clear_2, modulus);
-            let d0: i64 = cks.decrypt_signed(&ctxt_0);
-            assert_eq!(d0, clear_lhs, "Failed sanity decryption check");
-
-            let (ct_res, result_overflowed) = executor.execute((&ctxt_0, clear_rhs));
-            assert!(ct_res.block_carries_are_empty());
-            let (expected_result, expected_overflowed) =
-                signed_overflowing_sub_under_modulus(clear_lhs, clear_rhs, modulus);
-
-            let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-            let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-            assert_eq!(
-                decrypted_result, expected_result,
-                "Invalid result for sub, for ({clear_lhs} + {clear_rhs}) % {modulus} \
-                expected {expected_result}, got {decrypted_result}"
-            );
-            assert_eq!(
-                decrypted_overflowed,
-                expected_overflowed,
-                "Invalid overflow flag result for overflowing_sub, for ({clear_lhs} - {clear_rhs}) % {modulus} \
-                expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-            );
-            assert_eq!(result_overflowed.0.degree.get(), 1);
-            assert_eq!(result_overflowed.0.noise_level(), NoiseLevel::NOMINAL);
-        }
-    }
-
-    // Test with trivial inputs
-    for _ in 0..4 {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let a: SignedRadixCiphertext = sks.create_trivial_radix(clear_0, NB_CTXT);
-
-        let (encrypted_result, encrypted_overflow) = executor.execute((&a, clear_1));
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_sub_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&encrypted_result);
-        let decrypted_overflowed = cks.decrypt_bool(&encrypted_overflow);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for add, for ({clear_0} - {clear_1}) % {modulus} \
-                expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_sub, for ({clear_0} - {clear_1}) % {modulus} \
-                expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert_eq!(encrypted_overflow.0.degree.get(), 1);
-        assert_eq!(encrypted_overflow.0.noise_level(), NoiseLevel::ZERO);
-    }
-
-    // Test with scalar that is bigger than ciphertext modulus
-    for _ in 0..2 {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen_range(modulus..=i64::MAX);
-
-        let a = cks.encrypt_signed(clear_0);
-
-        let (encrypted_result, encrypted_overflow) = executor.execute((&a, clear_1));
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_sub_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&encrypted_result);
-        let decrypted_overflowed = cks.decrypt_bool(&encrypted_overflow);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for sub, for ({clear_0} - {clear_1}) % {modulus} \
-                expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_sub, for ({clear_0} - {clear_1}) % {modulus} \
-                expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert!(decrypted_overflowed); // Actually we know its an overflow case
-        assert_eq!(encrypted_overflow.0.degree.get(), 1);
-        assert_eq!(encrypted_overflow.0.noise_level(), NoiseLevel::ZERO);
-    }
-}
-
-pub(crate) fn signed_unchecked_bitand_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    for (clear_0, clear_1) in
-        create_iterator_of_signed_random_pairs::<NB_TESTS_UNCHECKED>(&mut rng, modulus)
-    {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = clear_0 & clear_1;
-        assert_eq!(clear_res, dec_res);
-    }
-}
-
-pub(crate) fn signed_unchecked_bitor_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    for (clear_0, clear_1) in
-        create_iterator_of_signed_random_pairs::<NB_TESTS_UNCHECKED>(&mut rng, modulus)
-    {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = clear_0 | clear_1;
-        assert_eq!(clear_res, dec_res);
-    }
-}
-
-pub(crate) fn signed_unchecked_bitxor_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    for (clear_0, clear_1) in
-        create_iterator_of_signed_random_pairs::<NB_TESTS_UNCHECKED>(&mut rng, modulus)
-    {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = clear_0 ^ clear_1;
-        assert_eq!(clear_res, dec_res);
-    }
-}
-
-pub(crate) fn signed_default_bitnot_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<&'a SignedRadixCiphertext, SignedRadixCiphertext>,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    sks.set_deterministic_pbs_execution(true);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks);
-
-    for _ in 0..NB_TESTS {
-        let clear_0 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let ct_res = executor.execute(&ctxt_0);
-        let ct_res2 = executor.execute(&ctxt_0);
-        assert_eq!(ct_res, ct_res2);
-
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = !clear_0;
-        assert_eq!(clear_res, dec_res);
-    }
-}
-
-pub(crate) fn signed_default_bitand_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    sks.set_deterministic_pbs_execution(true);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks.clone());
-
-    for _ in 0..NB_TESTS {
-        let mut clear_0 = rng.gen::<i64>() % modulus;
-        let mut clear_1 = rng.gen::<i64>() % modulus;
-
-        let mut ctxt_0 = cks.encrypt_signed(clear_0);
-        let mut ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let ct_res2 = executor.execute((&ctxt_0, &ctxt_1));
-        assert_eq!(ct_res, ct_res2);
-
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = clear_0 & clear_1;
-        assert_eq!(clear_res, dec_res);
-
-        let clear_2 = random_non_zero_value(&mut rng, modulus);
-        let clear_3 = random_non_zero_value(&mut rng, modulus);
-
-        sks.unchecked_scalar_add_assign(&mut ctxt_0, clear_2);
-        sks.unchecked_scalar_add_assign(&mut ctxt_1, clear_3);
-
-        assert!(!ctxt_0.block_carries_are_empty());
-        assert!(!ctxt_1.block_carries_are_empty());
-
-        clear_0 = signed_add_under_modulus(clear_0, clear_2, modulus);
-        clear_1 = signed_add_under_modulus(clear_1, clear_3, modulus);
-
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        assert!(ct_res.block_carries_are_empty());
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-
-        let expected_result = clear_0 & clear_1;
-        assert_eq!(dec_res, expected_result);
-    }
-}
-
-pub(crate) fn signed_default_bitor_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    sks.set_deterministic_pbs_execution(true);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks.clone());
-
-    for _ in 0..NB_TESTS {
-        let mut clear_0 = rng.gen::<i64>() % modulus;
-        let mut clear_1 = rng.gen::<i64>() % modulus;
-
-        let mut ctxt_0 = cks.encrypt_signed(clear_0);
-        let mut ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let ct_res2 = executor.execute((&ctxt_0, &ctxt_1));
-        assert_eq!(ct_res, ct_res2);
-
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = clear_0 | clear_1;
-        assert_eq!(clear_res, dec_res);
-
-        let clear_2 = random_non_zero_value(&mut rng, modulus);
-        let clear_3 = random_non_zero_value(&mut rng, modulus);
-
-        sks.unchecked_scalar_add_assign(&mut ctxt_0, clear_2);
-        sks.unchecked_scalar_add_assign(&mut ctxt_1, clear_3);
-
-        assert!(!ctxt_0.block_carries_are_empty());
-        assert!(!ctxt_1.block_carries_are_empty());
-
-        clear_0 = signed_add_under_modulus(clear_0, clear_2, modulus);
-        clear_1 = signed_add_under_modulus(clear_1, clear_3, modulus);
-
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        assert!(ct_res.block_carries_are_empty());
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-
-        let expected_result = clear_0 | clear_1;
-        assert_eq!(dec_res, expected_result);
-    }
-}
-
-pub(crate) fn signed_default_bitxor_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<
-        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
-        SignedRadixCiphertext,
-    >,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    sks.set_deterministic_pbs_execution(true);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks.clone());
-
-    for _ in 0..NB_TESTS {
-        let mut clear_0 = rng.gen::<i64>() % modulus;
-        let mut clear_1 = rng.gen::<i64>() % modulus;
-
-        let mut ctxt_0 = cks.encrypt_signed(clear_0);
-        let mut ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        let ct_res2 = executor.execute((&ctxt_0, &ctxt_1));
-        assert_eq!(ct_res, ct_res2);
-
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = clear_0 ^ clear_1;
-        assert_eq!(clear_res, dec_res);
-
-        let clear_2 = random_non_zero_value(&mut rng, modulus);
-        let clear_3 = random_non_zero_value(&mut rng, modulus);
-
-        sks.unchecked_scalar_add_assign(&mut ctxt_0, clear_2);
-        sks.unchecked_scalar_add_assign(&mut ctxt_1, clear_3);
-
-        assert!(!ctxt_0.block_carries_are_empty());
-        assert!(!ctxt_1.block_carries_are_empty());
-
-        clear_0 = signed_add_under_modulus(clear_0, clear_2, modulus);
-        clear_1 = signed_add_under_modulus(clear_1, clear_3, modulus);
-
-        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
-        assert!(ct_res.block_carries_are_empty());
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-
-        let expected_result = clear_0 ^ clear_1;
-        assert_eq!(dec_res, expected_result);
-    }
-}
-
-pub(crate) fn signed_default_scalar_bitand_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<(&'a SignedRadixCiphertext, i64), SignedRadixCiphertext>,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    sks.set_deterministic_pbs_execution(true);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks.clone());
-
-    for _ in 0..NB_TESTS {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let mut ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let ct_res = executor.execute((&ctxt_0, clear_1));
-        let ct_res2 = executor.execute((&ctxt_0, clear_1));
-        assert_eq!(ct_res, ct_res2);
-
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = clear_0 & clear_1;
-        assert_eq!(clear_res, dec_res);
-
-        let clear_2 = random_non_zero_value(&mut rng, modulus);
-        sks.unchecked_scalar_add_assign(&mut ctxt_0, clear_2);
-        assert!(!ctxt_0.block_carries_are_empty());
-
-        let ct_res = executor.execute((&ctxt_0, clear_1));
-        assert!(ct_res.block_carries_are_empty());
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-
-        let expected_result = signed_add_under_modulus(clear_0, clear_2, modulus) & clear_1;
-        assert_eq!(dec_res, expected_result);
-    }
-}
-
-pub(crate) fn signed_default_scalar_bitor_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<(&'a SignedRadixCiphertext, i64), SignedRadixCiphertext>,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    sks.set_deterministic_pbs_execution(true);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks.clone());
-
-    for _ in 0..NB_TESTS {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let mut ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let ct_res = executor.execute((&ctxt_0, clear_1));
-        let ct_res2 = executor.execute((&ctxt_0, clear_1));
-        assert_eq!(ct_res, ct_res2);
-
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = clear_0 | clear_1;
-        assert_eq!(clear_res, dec_res);
-
-        let clear_2 = random_non_zero_value(&mut rng, modulus);
-
-        sks.unchecked_scalar_add_assign(&mut ctxt_0, clear_2);
-        assert!(!ctxt_0.block_carries_are_empty());
-
-        let ct_res = executor.execute((&ctxt_0, clear_1));
-        assert!(ct_res.block_carries_are_empty());
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-
-        let expected_result = signed_add_under_modulus(clear_0, clear_2, modulus) | clear_1;
-        assert_eq!(dec_res, expected_result);
-    }
-}
-
-pub(crate) fn signed_default_scalar_bitxor_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<(&'a SignedRadixCiphertext, i64), SignedRadixCiphertext>,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    sks.set_deterministic_pbs_execution(true);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    executor.setup(&cks, sks.clone());
-
-    for _ in 0..NB_TESTS {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let mut ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let ct_res = executor.execute((&ctxt_0, clear_1));
-        let ct_res2 = executor.execute((&ctxt_0, clear_1));
-        assert_eq!(ct_res, ct_res2);
-
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-        let clear_res = clear_0 ^ clear_1;
-        assert_eq!(clear_res, dec_res);
-
-        let clear_2 = random_non_zero_value(&mut rng, modulus);
-
-        sks.unchecked_scalar_add_assign(&mut ctxt_0, clear_2);
-        assert!(!ctxt_0.block_carries_are_empty());
-
-        let ct_res = executor.execute((&ctxt_0, clear_1));
-        assert!(ct_res.block_carries_are_empty());
-        let dec_res: i64 = cks.decrypt_signed(&ct_res);
-
-        let expected_result = signed_add_under_modulus(clear_0, clear_2, modulus) ^ clear_1;
-        assert_eq!(dec_res, expected_result);
-    }
-}
-
-pub(crate) fn signed_unchecked_scalar_left_shift_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<(&'a SignedRadixCiphertext, i64), SignedRadixCiphertext>,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    executor.setup(&cks, sks);
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-    assert!(modulus > 0);
-    assert!((modulus as u64).is_power_of_two());
-    let nb_bits = modulus.ilog2() + 1; // We are using signed numbers
-
-    for _ in 0..NB_TESTS {
-        let clear = rng.gen::<i64>() % modulus;
-        let clear_shift = rng.gen::<u32>();
-
-        let ct = cks.encrypt_signed(clear);
-
-        // case when 0 <= shift < nb_bits
-        {
-            let clear_shift = clear_shift % nb_bits;
-            let ct_res = executor.execute((&ct, clear_shift as i64));
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-            let expected = signed_left_shift_under_modulus(clear, clear_shift, modulus);
-            assert_eq!(expected, dec_res);
-        }
-
-        // case when shift >= nb_bits
-        {
-            let clear_shift = clear_shift.saturating_add(nb_bits);
-            let ct_res = executor.execute((&ct, clear_shift as i64));
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-            let expected = signed_left_shift_under_modulus(clear, clear_shift % nb_bits, modulus);
-            assert_eq!(expected, dec_res);
-        }
-    }
-}
-
-pub(crate) fn signed_unchecked_scalar_right_shift_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<(&'a SignedRadixCiphertext, i64), SignedRadixCiphertext>,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    executor.setup(&cks, sks);
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-    assert!(modulus > 0);
-    assert!((modulus as u64).is_power_of_two());
-    let nb_bits = modulus.ilog2() + 1; // We are using signed numbers
-
-    for _ in 0..NB_TESTS {
-        let clear = rng.gen::<i64>() % modulus;
-        let clear_shift = rng.gen::<u32>();
-
-        let ct = cks.encrypt_signed(clear);
-
-        // case when 0 <= shift < nb_bits
-        {
-            let clear_shift = clear_shift % nb_bits;
-            let ct_res = executor.execute((&ct, clear_shift as i64));
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-            let expected = signed_right_shift_under_modulus(clear, clear_shift, modulus);
-            assert_eq!(expected, dec_res);
-        }
-
-        // case when shift >= nb_bits
-        {
-            let clear_shift = clear_shift.saturating_add(nb_bits);
-            let ct_res = executor.execute((&ct, clear_shift as i64));
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-            let expected = signed_right_shift_under_modulus(clear, clear_shift % nb_bits, modulus);
-            assert_eq!(expected, dec_res);
-        }
-    }
-}
-
-pub(crate) fn signed_default_scalar_left_shift_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<(&'a SignedRadixCiphertext, i64), SignedRadixCiphertext>,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    sks.set_deterministic_pbs_execution(true);
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    executor.setup(&cks, sks.clone());
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    assert!(modulus > 0);
-    assert!((modulus as u64).is_power_of_two());
-    let nb_bits = modulus.ilog2() + 1; // We are using signed numbers
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let mut clear = rng.gen::<i64>() % modulus;
-
-        let offset = random_non_zero_value(&mut rng, modulus);
-
-        let mut ct = cks.encrypt_signed(clear);
-        sks.unchecked_scalar_add_assign(&mut ct, offset);
-        clear = signed_add_under_modulus(clear, offset, modulus);
-
-        // case when 0 <= shift < nb_bits
-        {
-            let clear_shift = rng.gen::<u32>() % nb_bits;
-            let ct_res = executor.execute((&ct, clear_shift as i64));
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-            let clear_res = signed_left_shift_under_modulus(clear, clear_shift, modulus);
-            assert_eq!(
-                clear_res, dec_res,
-                "Invalid left shift result, for '{clear} << {clear_shift}', \
-                expected:  {clear_res}, got: {dec_res}"
-            );
-
-            let ct_res2 = executor.execute((&ct, clear_shift as i64));
-            assert_eq!(ct_res, ct_res2, "Failed determinism check");
-        }
-
-        // case when shift >= nb_bits
-        {
-            let clear_shift = rng.gen_range(nb_bits..=u32::MAX);
-            let ct_res = executor.execute((&ct, clear_shift as i64));
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-            // We mimic wrapping_shl manually as we use a bigger type
-            // than the nb_bits we actually simulate in this test
-            let clear_res = signed_left_shift_under_modulus(clear, clear_shift % nb_bits, modulus);
-            assert_eq!(
-                clear_res, dec_res,
-                "Invalid left shift result, for '{clear} << {clear_shift}', \
-                expected:  {clear_res}, got: {dec_res}"
-            );
-
-            let ct_res2 = executor.execute((&ct, clear_shift as i64));
-            assert_eq!(ct_res, ct_res2, "Failed determinism check");
-        }
-    }
-}
-
-pub(crate) fn signed_default_scalar_right_shift_test<P, T>(param: P, mut executor: T)
-where
-    P: Into<PBSParameters>,
-    T: for<'a> FunctionExecutor<(&'a SignedRadixCiphertext, i64), SignedRadixCiphertext>,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-    sks.set_deterministic_pbs_execution(true);
-    let sks = Arc::new(sks);
-
-    let mut rng = rand::thread_rng();
-
-    executor.setup(&cks, sks.clone());
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    assert!(modulus > 0);
-    assert!((modulus as u64).is_power_of_two());
-    let nb_bits = modulus.ilog2() + 1; // We are using signed numbers
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let mut clear = rng.gen::<i64>() % modulus;
-
-        let offset = random_non_zero_value(&mut rng, modulus);
-
-        let mut ct = cks.encrypt_signed(clear);
-        sks.unchecked_scalar_add_assign(&mut ct, offset);
-        clear = signed_add_under_modulus(clear, offset, modulus);
-
-        // case when 0 <= shift < nb_bits
-        {
-            let clear_shift = rng.gen::<u32>() % nb_bits;
-            let ct_res = executor.execute((&ct, clear_shift as i64));
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-            let clear_res = signed_right_shift_under_modulus(clear, clear_shift, modulus);
-            assert_eq!(
-                clear_res, dec_res,
-                "Invalid right shift result, for '{clear} >> {clear_shift}', \
-                expected:  {clear_res}, got: {dec_res}"
-            );
-
-            let ct_res2 = executor.execute((&ct, clear_shift as i64));
-            assert_eq!(ct_res, ct_res2, "Failed determinism check");
-        }
-
-        // case when shift >= nb_bits
-        {
-            let clear_shift = rng.gen_range(nb_bits..=u32::MAX);
-            let ct_res = executor.execute((&ct, clear_shift as i64));
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-            // We mimic wrapping_shl manually as we use a bigger type
-            // than the nb_bits we actually simulate in this test
-            let clear_res = signed_right_shift_under_modulus(clear, clear_shift % nb_bits, modulus);
-            assert_eq!(
-                clear_res, dec_res,
-                "Invalid right shift result, for '{clear} >> {clear_shift}', \
-                expected:  {clear_res}, got: {dec_res}"
-            );
-
-            let ct_res2 = executor.execute((&ct, clear_shift as i64));
-            assert_eq!(ct_res, ct_res2, "Failed determinism check");
-        }
-    }
-}
