+++ tfhe/src/integer/server_key/radix_parallel/tests_signed/test_add.rs
@@ -1,11 +1,9 @@
 use crate::integer::keycache::KEY_CACHE;
 use crate::integer::server_key::radix_parallel::sub::SignedOperation;
-use crate::integer::server_key::radix_parallel::tests_cases_signed::{
-    random_non_zero_value, signed_add_under_modulus, signed_default_add_test,
-    signed_overflowing_add_under_modulus, signed_smart_add_test, signed_unchecked_add_test,
-};
+use crate::integer::server_key::radix_parallel::tests_cases_unsigned::FunctionExecutor;
 use crate::integer::server_key::radix_parallel::tests_signed::{
-    NB_CTXT, NB_TESTS, NB_TESTS_SMALLER,
+    create_iterator_of_signed_random_pairs, random_non_zero_value, signed_add_under_modulus,
+    signed_overflowing_add_under_modulus, NB_CTXT, NB_TESTS, NB_TESTS_SMALLER, NB_TESTS_UNCHECKED,
 };
 use crate::integer::server_key::radix_parallel::tests_unsigned::CpuFunctionExecutor;
 use crate::integer::tests::create_parametrized_test;
@@ -17,6 +15,7 @@ use crate::shortint::ciphertext::NoiseLevel;
 use crate::shortint::parameters::coverage_parameters::*;
 use crate::shortint::parameters::*;
 use rand::Rng;
+use std::sync::Arc;
 
 create_parametrized_test!(integer_signed_unchecked_add);
 create_parametrized_test!(integer_signed_unchecked_overflowing_add);
@@ -331,4 +330,147 @@ where
 {
     let executor = CpuFunctionExecutor::new(&ServerKey::smart_add_parallelized);
     signed_smart_add_test(param, executor);
+}
+
+pub(crate) fn signed_unchecked_add_test<P, T>(param: P, mut executor: T)
+where
+    P: Into<PBSParameters>,
+    T: for<'a> FunctionExecutor<
+        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
+        SignedRadixCiphertext,
+    >,
+{
+    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
+    let sks = Arc::new(sks);
+    let cks = RadixClientKey::from((cks, NB_CTXT));
+
+    let mut rng = rand::thread_rng();
+
+    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
+
+    executor.setup(&cks, sks);
+
+    // check some overflow behaviour
+    let overflowing_values = [
+        (-modulus, -1, modulus - 1),
+        (modulus - 1, 1, -modulus),
+        (-modulus, -2, modulus - 2),
+        (modulus - 2, 2, -modulus),
+    ];
+    for (clear_0, clear_1, expected_clear) in overflowing_values {
+        let ctxt_0 = cks.encrypt_signed(clear_0);
+        let ctxt_1 = cks.encrypt_signed(clear_1);
+        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
+        let dec_res: i64 = cks.decrypt_signed(&ct_res);
+        let clear_res = signed_add_under_modulus(clear_0, clear_1, modulus);
+        assert_eq!(clear_res, dec_res);
+        assert_eq!(clear_res, expected_clear);
+    }
+
+    for (clear_0, clear_1) in
+        create_iterator_of_signed_random_pairs::<{ NB_TESTS_UNCHECKED }>(&mut rng, modulus)
+    {
+        let ctxt_0 = cks.encrypt_signed(clear_0);
+        let ctxt_1 = cks.encrypt_signed(clear_1);
+
+        let ct_res = executor.execute((&ctxt_0, &ctxt_1));
+        let dec_res: i64 = cks.decrypt_signed(&ct_res);
+        let clear_res = signed_add_under_modulus(clear_0, clear_1, modulus);
+        assert_eq!(clear_res, dec_res);
+    }
+}
+
+pub(crate) fn signed_default_add_test<P, T>(param: P, mut executor: T)
+where
+    P: Into<PBSParameters>,
+    T: for<'a> FunctionExecutor<
+        (&'a SignedRadixCiphertext, &'a SignedRadixCiphertext),
+        SignedRadixCiphertext,
+    >,
+{
+    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
+    sks.set_deterministic_pbs_execution(true);
+    let sks = Arc::new(sks);
+    let cks = RadixClientKey::from((cks, NB_CTXT));
+
+    let mut rng = rand::thread_rng();
+
+    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
+
+    executor.setup(&cks, sks);
+
+    let mut clear;
+
+    for _ in 0..NB_TESTS_SMALLER {
+        let clear_0 = rng.gen::<i64>() % modulus;
+        let clear_1 = rng.gen::<i64>() % modulus;
+
+        let ctxt_0 = cks.encrypt_signed(clear_0);
+        let ctxt_1 = cks.encrypt_signed(clear_1);
+
+        let mut ct_res = executor.execute((&ctxt_0, &ctxt_1));
+        let tmp_ct = executor.execute((&ctxt_0, &ctxt_1));
+        assert!(ct_res.block_carries_are_empty());
+        assert_eq!(ct_res, tmp_ct);
+
+        clear = signed_add_under_modulus(clear_0, clear_1, modulus);
+
+        // println!("clear_0 = {}, clear_1 = {}", clear_0, clear_1);
+        // add multiple times to raise the degree
+        for _ in 0..NB_TESTS_SMALLER {
+            ct_res = executor.execute((&ct_res, &ctxt_0));
+            assert!(ct_res.block_carries_are_empty());
+            clear = signed_add_under_modulus(clear, clear_0, modulus);
+
+            let dec_res: i64 = cks.decrypt_signed(&ct_res);
+
+            // println!("clear = {}, dec_res = {}", clear, dec_res);
+            assert_eq!(clear, dec_res);
+        }
+    }
+}
+
+pub(crate) fn signed_smart_add_test<P, T>(param: P, mut executor: T)
+where
+    P: Into<PBSParameters>,
+    T: for<'a> FunctionExecutor<
+        (&'a mut SignedRadixCiphertext, &'a mut SignedRadixCiphertext),
+        SignedRadixCiphertext,
+    >,
+{
+    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
+    let sks = Arc::new(sks);
+    let cks = RadixClientKey::from((cks, NB_CTXT));
+
+    let mut rng = rand::thread_rng();
+
+    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
+
+    executor.setup(&cks, sks);
+
+    let mut clear;
+
+    for _ in 0..NB_TESTS_SMALLER {
+        let clear_0 = rng.gen_range(-modulus..modulus);
+        let clear_1 = rng.gen_range(-modulus..modulus);
+
+        let mut ctxt_0 = cks.encrypt_signed(clear_0);
+        let mut ctxt_1 = cks.encrypt_signed(clear_1);
+
+        let mut ct_res = executor.execute((&mut ctxt_0, &mut ctxt_1));
+        clear = signed_add_under_modulus(clear_0, clear_1, modulus);
+        let dec_res: i64 = cks.decrypt_signed(&ct_res);
+        assert_eq!(clear, dec_res);
+
+        // add multiple times to raise the degree
+        for _ in 0..NB_TESTS_SMALLER {
+            ct_res = executor.execute((&mut ct_res, &mut ctxt_0));
+            clear = signed_add_under_modulus(clear, clear_0, modulus);
+
+            let dec_res: i64 = cks.decrypt_signed(&ct_res);
+
+            // println!("clear = {}, dec_res = {}", clear, dec_res);
+            assert_eq!(clear, dec_res);
+        }
+    }
 }
