+++ tfhe/src/integer/server_key/radix_parallel/tests_signed/test_rotate.rs
@@ -1,9 +1,7 @@
 use crate::integer::keycache::KEY_CACHE;
-use crate::integer::server_key::radix_parallel::tests_cases_signed::{
-    random_non_zero_value, rotate_left_helper, rotate_right_helper, signed_add_under_modulus,
-};
 use crate::integer::server_key::radix_parallel::tests_cases_unsigned::FunctionExecutor;
 use crate::integer::server_key::radix_parallel::tests_signed::{
+    random_non_zero_value, rotate_left_helper, rotate_right_helper, signed_add_under_modulus,
     NB_CTXT, NB_TESTS, NB_TESTS_SMALLER,
 };
 use crate::integer::server_key::radix_parallel::tests_unsigned::CpuFunctionExecutor;
