+++ tfhe/src/integer/server_key/radix_parallel/tests_signed/test_shift.rs
@@ -1,11 +1,8 @@
 use crate::integer::keycache::KEY_CACHE;
-use crate::integer::server_key::radix_parallel::tests_cases_signed::{
-    random_non_zero_value, signed_add_under_modulus, signed_left_shift_under_modulus,
-    signed_right_shift_under_modulus,
-};
 use crate::integer::server_key::radix_parallel::tests_cases_unsigned::FunctionExecutor;
 use crate::integer::server_key::radix_parallel::tests_signed::{
-    NB_CTXT, NB_TESTS, NB_TESTS_SMALLER,
+    random_non_zero_value, signed_add_under_modulus, signed_left_shift_under_modulus,
+    signed_right_shift_under_modulus, NB_CTXT, NB_TESTS, NB_TESTS_SMALLER,
 };
 use crate::integer::server_key::radix_parallel::tests_unsigned::CpuFunctionExecutor;
 use crate::integer::tests::create_parametrized_test;
